requirements.txt
fastapi[all]
uvicorn
requests
google-generativeai>=0.3.1
openai>=1.0.0
python-dotenv
google-auth>=2.0.0
google-auth-oauthlib
google-api-python-client
faiss-cpu
tiktoken
langchain==0.2.0
langchain-community
langchain-openai
httpx
python-dotenv